hpav_test: use proper exit codes
When using the tool in e.g. scripts, it is often desirable to know whether
the execution of the tool was successful or whether an error occurred.

This commit reworks the exit code generation and propagation through
several code paths.

The goal and behavior was also documented in the README file for users.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>