Big refactoring: factor out common code and tidy up
Many functions use the very same code which makes it difficult
to maintain and extend the tool.
This change introduces new functions which are then integrated
into existing functionality.

The original idea was to replace the interface number handling
with code that allows to use existing interface _names_ (at least
on Linux). Reason is that using interface _numbers_ requires two
steps in shell scripts: with the first step, you have to determine
the interface number and in the second step, you can pass it
to the actual hpav_test command you want to execute. However, this
is racy, especially during the boot phase of a Linux system, when
interfaces appear and the interface numbering may change.
This could result in sending/expecting the chip on the wrong
interface.
So the initial idea was to extend the tool that both, a number
or a name, can be passed as argument.
While implementing, it then turned out, that it makes sense
to tune the whole code base a little bit.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>